Lab3/requirements.txt
flask==2.3.3
kafka-python==2.0.2
psycopg2-binary==2.9.9
pandas==2.0.3
python-dotenv==1.0.0
